fix: layer header layout
