+++ sites/geohub/src/components/pages/map/layers/header/VisibilityButton.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import type { Layer } from '$lib/types';
-	import { layerList, MAPSTORE_CONTEXT_KEY, type MapStore } from '$stores';
+	import { MAPSTORE_CONTEXT_KEY, layerList, type MapStore } from '$stores';
 	import { cloneDeep } from 'lodash-es';
 	import { getContext } from 'svelte';
 
@@ -56,7 +56,7 @@
 		on:click={() => toggleVisibility()}
 		on:keydown={handleKeyDown}
 	>
-		<i class="fa-solid {visibility === 'visible' ? 'fa-eye' : 'fa-eye-slash'} fa-sm" />
+		<i class="fa-solid {visibility === 'visible' ? 'fa-eye' : 'fa-eye-slash'} fa-xl" />
 	</div>
 </div>
